requirements.txt
streamlit
pytesseract
Pillow
pdf2image
